+++ src/__tests__/DIContainer.test.ts
@@ -1,4 +1,4 @@
 import { test_IDIContainer } from './IDIContainer.spec';
-import { TSinjex } from '../TSinjex';
+import { TSinjex } from '../classes/TSinjex';
 
 test_IDIContainer(TSinjex);
