+++ src/decorators/Register.ts
@@ -1,12 +1,14 @@
-import { TSinjex } from '../TSinjex';
+import { Identifier } from 'src/types/Identifier';
+import { TSinjex } from '../classes/TSinjex';
 
 /**
- * A decorator to register a class in the DI (Dependency Injection) container.
+ * A decorator to register a class in the **TSinjex** DI (Dependency Injection) container.
  * @template TargetType The type of the class to be registered.
  * @param identifier The identifier used to register the class in the DI container.
- * @param deprecated If true, the dependency is deprecated => a warning
- * is logged when the dependency is resolved.
- * @returns A function that is applied as a decorator to the class.
+ * @see {@link Identifier} for more information on identifiers.
+ * @param deprecated If true, the dependency is deprecated and a warning
+ * is logged only once upon the first resolution of the dependency.
+ * @returns The decorator function to be applied on the class.
  * @example
  * ```ts
  * \@Register('MyClassIdentifier')
@@ -17,7 +19,7 @@ import { TSinjex } from '../TSinjex';
  */
 export function Register<
     TargetType extends new (...args: unknown[]) => InstanceType<TargetType>,
->(identifier: string, deprecated?: boolean) {
+>(identifier: Identifier, deprecated?: boolean) {
     return function (constructor: TargetType, ...args: unknown[]): void {
         // Get the instance of the DI container
         const diContainer = TSinjex.getInstance();
