+++ src/decorators/RegisterInstance.ts
@@ -1,17 +1,19 @@
-import { TSinjex } from '../TSinjex';
+import { Identifier } from 'src/types/Identifier';
+import { TSinjex } from '../classes/TSinjex';
 import { InitDelegate } from '../types/InitDelegate';
 
 /**
  * A decorator to register an instance of a class in the DI (Dependency Injection) container.
- * The instance is created only when it is first needed (Lazy Initialization).
  * @template TargetType The type of the class whose instance is to be registered.
  * @param identifier The identifier used to register the instance in the DI container.
+ * @see {@link Identifier} for more information on identifiers.
  * @param init An optional initializer function which get the constructor of the class
  * as input and returns an instance of the class.
- * @returns A function that is applied as a decorator to the class.
+ * @see {@link InitDelegate} for more information on initializer functions.
+ * @returns The decorator function to be applied on the class.
  * @example
  * ```ts
- * \@RegisterInstance('MyClassInstanceIdentifier', arg1, arg2)
+ * \@RegisterInstance('MyClassInstanceIdentifier', (constructor) => new constructor())
  * class MyClass {
  *   // ...
  * }
@@ -20,7 +22,7 @@ import { InitDelegate } from '../types/InitDelegate';
 export function RegisterInstance<
     TargetType extends new (..._args: unknown[]) => InstanceType<TargetType>,
 >(
-    identifier: string,
+    identifier: Identifier,
     init?: InitDelegate<
         TargetType & { new (..._args: unknown[]): InstanceType<TargetType> },
         InstanceType<TargetType>
